backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
sqlalchemy
psycopg2-binary
requests
openai